Fix binary validation (#17)
The update command tries to validate the downloaded dnvm binary to ensure that it can
do basic things. One thing it checks is whether it can print a help message which
contains the 'usage: ' string. However, previously dnvm printed the usage message to
stderr instead of stdout, so the validation failed. The handler has been updated to
print help to stdout iff it's requested.