Rework self-update (#55)
Currently self-update is something done entirely by the originating
binary -- it checks for available updates, downloads the updaate, and
swaps the new binary in for the old.

This means that the new binary doesn't have a chance to run any program
logic.

The new model is to download the new binary, and then run it in an
installer mode. This will let the binary install itself, running
whatever's necessary to do so.