Add more versions to the Manifest
In order to implement uninstall, we need versions for each of the
components that make up an SDK. Those versions were not previously
stored in the manifest, and not available from the APIs we were calling.
This PR changes the manifest to contain these versions and adds support
for pulling more detailed information from the index.