Install preview SDKs to the default folder (#111)
Previously, preview versions were installed to the 'preview' folder to
separate from the release versions. There are drawbacks to installing in
the same folder, but they can mostly be addressed by using global.json.
Using one folder is also preferred for the official product, so this matches
the expected product layout.