Fix installing daily builds
Normally we rely on the .NET releases index to find both the version
information for the requested SDK and the download URL. In the case of
daily builds, none of that information is published to an index.
Therefore we need to use other resources. For the URL, there is a known
pattern that is implemented by dotnet-install.sh that we copy. For the
version information, there is an undocumented `productCommit-{RID}.json`
file that appears to contain the necessary information. Together, these
items should allow us to install and uninstall daily builds.