Fix tracking channels with no builds
It's common for some channels (e.g., preview) to not have builds for a
period of time when the previous channel has gone out of support, but
the new one has not had any releases yet. Tracking shouldn't depend on
installing an SDK, so there's no reason we can't allow tracking to
proceed.

Fixes #135