Adopt CancelScope and ScopedHttpClient (#176)
This moves the repo to using a CancelScope instead of passing around
CancellationTokens. This is a design pattern from structured
concurrency. It also introduces a ScopedHttpClient which respects
CancelScope. The ScopedHttpClient is also attached to a DnvmEnv, instead
of being a global static, which allows for testing and more
configuration options.

Lastly, the ScopedHttpClient now no longer has a 100 second global
timeout. Instead, timeouts are regulated per-scope or through the
specialty GetStringAsync method (which has a default timeout).