fix: Use symlinks on Windows instead of cmd file (#197)
One of the core pieces of functionality in dnvm is swapping between
different isolated installs in different directories. To switch between
directories, the implementation on Unix-like systems uses symlinks. The
Windows implementation used a cmd file since the muxer didn't support
being run behind a symlink. In .NET 9 that restriction was lifted, so we
can use symlinks everywhere.

Fixes #189