Fix restoring previews and daily builds
There were two underlying bugs. The first was that `restore` wasn't
properly doing version comparisons for preview builds. The second was
that it didn't check the feed URL directly in the case that the index
was missing the build. Since daily builds are never in the index, this
meant you couldn't restore a daily build.

This PR fixes both issues and adds regression tests.