Move install config to an env variable
Rather than prompting the user for an output path, use
the DNVM_HOME env variable to decide the install path.
This is a relatively uncommon scneario and requiring an
env variable indicates how it will also be used in the
future: via setting the DNVM_HOME variable for all shells.