Only prune tracked SDKs
There are multiple ways to install SDKs:
* via a tracked channel
* via a manual install
* via restore

Prune is meant to remove SDKs that have been superseded.
However, the relationships between each install isn't clear,
except for installations into a tracked channel. In that case,
it's clear when a newer SDK takes priority over an old one. This
change causes prune to only remove SDKs that were installed as
part of a tracked channel where the channel now contains a newer
version.